feat: 메인 클래스에 있던 빈 설정 별도 파일로 분리
- 메인 클래스 이름 변경
- 불필요한 설정 애노테이션 제거
