+++ src/main/java/dailyquest/config/JpaAuditingConfiguration.java
@@ -1,9 +1,25 @@
 package dailyquest.config;
 
+import dailyquest.user.dto.UserPrincipal;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+
+import java.util.Optional;
 
 @Configuration
 @EnableJpaAuditing
 public class JpaAuditingConfiguration {
+    @Bean
+    public AuditorAware<Long> auditorProvider() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (null == authentication || !authentication.isAuthenticated()) {
+            return null;
+        }
+        UserPrincipal user = (UserPrincipal) authentication.getPrincipal();
+        return () -> Optional.of(user.getId());
+    }
 }
